+++ infra/migrations/0007_living_betty_brant.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `collection` ADD `featured_on_landing` integer DEFAULT false NOT NULL;--> statement-breakpoint
+ALTER TABLE `collection` ADD `landing_position` integer DEFAULT 0 NOT NULL;